Move common types between server and client into "common" (#17)
