+++ common/bus.ts
@@ -10,4 +10,3 @@ export interface Bus {
     routeId: string;
   };
 }
-
