+++ common/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "real-time-bus-common-types",
+  "private": true
+}
