+++ common/point.ts
