+++ common/shape.ts
