+++ server/src/ingestion/gtfsRtIngestion.ts
@@ -8,6 +8,8 @@ export interface RawBusPosition {
   bus?: { id?: string; label?: string };
 }
 
+export interface BusPosition extends RawBusPosition {}
+
 export interface RawBusEntity {
   id: string;
   bus: RawBusPosition;
